add Arrival creatures from server

+++ server/src/main/kotlin/ballad/server/game/Step.kt
@@ -5,5 +5,5 @@ data class Step(
     val fromY: Int,
     val direction: Direction,
     val speed: Int,
-    val creatureId: Long
+    val creatureId: Int
 ) : Action
